ai-service/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.18.4
amplitude-analytics==1.2.2
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
apprise==1.9.9
asgi-lifespan==2.1.0
asyncpg==0.31.0
attrs==26.1.0
beartype==0.22.9
beautifulsoup4==4.14.3
billiard==4.2.4
blis==1.3.3
boto3==1.42.73
botocore==1.42.73
cachetools==7.0.5
catalogue==2.0.10
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.23.0
cloudpickle==3.1.2
colorama==0.4.6
confection==1.3.3
coolname==4.1.0
cronsim==2.7
cryptography==46.0.5
cyclopts==4.10.0
cymem==2.0.13
dataclasses-json==0.6.7
dateparser==1.3.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.4
emoji==2.15.0
exceptiongroup==1.3.1
fakeredis==2.34.1
fastapi==0.135.1
filelock==3.25.2
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.2.0
google-auth==2.49.1
google-genai==1.71.0
graphviz==0.21
greenlet==3.3.2
griffe==2.0.0
griffecli==2.0.0
griffelib==2.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
installer==0.7.0
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jiter==0.13.0
jmespath==1.1.0
jsonpatch==1.33
jsonpointer==3.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.28
langchain-google-genai==4.2.1
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langdetect==1.0.9
langsmith==0.7.22
llvmlite==0.47.0
lupa==2.6
lxml==6.0.2
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
numba==0.65.0
numpy==2.4.3
oauthlib==3.3.1
olefile==0.47
openai==2.31.0
opentelemetry-api==1.40.0
orjson==3.11.7
packaging==26.0
pandas==3.0.2
pathspec==1.0.4
pendulum==3.2.0
pgvector==0.4.2
pluggy==1.6.0
prefect==3.6.23
preshed==3.0.13
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
psycopg2-binary==2.9.11
py-key-value-aio==0.4.4
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydocket==0.18.2
Pygments==2.19.2
PyMuPDF==1.27.2.2
pypdf==6.9.2
pypdfium2==5.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-iso639==2026.1.31
python-json-logger==4.0.0
python-magic==0.4.27
python-multipart==0.0.24
python-oxmsg==0.0.2
python-slugify==8.0.4
pytz==2026.1.post1
PyYAML==6.0.3
RapidFuzz==3.14.5
readchar==4.2.1
redis==7.3.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
ruamel.yaml==0.19.1
ruamel.yaml.clib==0.2.15
s3transfer==0.16.0
scalar_fastapi==1.8.1
semver==3.0.4
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smart_open==7.5.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.48
srsly==2.5.3
starlette==1.0.0
tenacity==9.1.4
text-unidecode==1.3
thinc==8.3.13
tiktoken==0.12.0
toml==0.10.2
tqdm==4.67.3
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uncalled-for==0.2.0
unstructured==0.22.18
unstructured-client==0.43.2
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.6.0
weasel==1.0.0
webencodings==0.5.1
websockets==16.0
wrapt==2.1.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
